1.8: Fix registration for PlayerListeners
Sadly, Attribute isn't in the bukkit API for 1.8.x, Misat will find a new workaround to calculate the damage values for older versions
