+++ plugin/build.gradle
@@ -16,7 +16,7 @@ dependencies {
     }
     compileOnly 'io.netty:netty-all:4.1.42.Final'
     compileOnly 'net.citizensnpcs:citizens-main:2.0.27-SNAPSHOT'
-    compileOnly 'com.alessiodp.parties:parties-api:3.1.12'
+    compileOnly 'com.alessiodp.parties:parties-api:3.2.6'
 
     /* SHADED */
     implementation project(':BedWars-API')
