+++ plugin/src/main/java/org/screamingsandals/bedwars/listener/Player19Listener.java
@@ -19,6 +19,9 @@
 
 package org.screamingsandals.bedwars.listener;
 
+import org.bukkit.attribute.Attribute;
+import org.bukkit.attribute.AttributeInstance;
+import org.bukkit.event.player.PlayerMoveEvent;
 import org.screamingsandals.bedwars.Main;
 import org.screamingsandals.bedwars.api.game.GameStatus;
 import org.screamingsandals.bedwars.game.Game;
@@ -29,6 +32,8 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.ItemMergeEvent;
 import org.bukkit.event.player.PlayerSwapHandItemsEvent;
+
+import java.util.Arrays;
 
 public class Player19Listener implements Listener {
     @EventHandler
@@ -55,6 +60,44 @@ public class Player19Listener implements Listener {
                 if (GameCreator.isInArea(event.getEntity().getLocation(), game.getPos1(), game.getPos2()) || GameCreator.isInArea(event.getTarget().getLocation(), game.getPos1(), game.getPos2())) {
                     event.setCancelled(true);
                     return;
+                }
+            }
+        }
+    }
+
+    @EventHandler
+    public void onMove(PlayerMoveEvent event) {
+        if (event.isCancelled()) {
+            return;
+        }
+
+        Player player = event.getPlayer();
+        if (Main.isPlayerInGame(player)) {
+            GamePlayer gPlayer = Main.getPlayerGameProfile(player);
+            Game game = gPlayer.getGame();
+            if (game.getOriginalOrInheritedDamageWhenPlayerIsNotInArena() && game.getStatus() == GameStatus.RUNNING
+                    && !gPlayer.isSpectator) {
+                if (!GameCreator.isInArea(event.getTo(), game.getPos1(), game.getPos2())) {
+                    AttributeInstance armor = player.getAttribute(Attribute.GENERIC_ARMOR);
+                    AttributeInstance armorToughness = null;
+                    if (Arrays.stream(Attribute.values()).anyMatch(attribute -> "GENERIC_ARMOR_TOUGHNESS".equals(attribute.name()))) {
+                        armorToughness = player.getAttribute(Attribute.GENERIC_ARMOR_TOUGHNESS);
+                    }
+                    if (armor == null) {
+                        player.damage(5);
+                    } else {
+                        // this is not 100% accurate formula - armorToughness check contains weaponDamage which is hardcoded to 5 (4*5=20) but we don't know the weapon damage yet
+                        double multiplier = (1.0 - Math.min(20.0, Math.max(armor.getValue() / 5.0, armor.getValue() - 20.0 / ((armorToughness != null ? armorToughness.getValue() : 0) + 8))) / 25.0);
+                        if (multiplier < 1) {
+                            multiplier = 2 - multiplier;
+                        }
+                        double weaponDamage = multiplier * 5.0;
+                        player.damage(weaponDamage);
+                    }
+                }
+            } else if (Main.getConfigurator().config.getBoolean("preventSpectatorFlyingAway", false) && gPlayer.isSpectator && (game.getStatus() == GameStatus.RUNNING || game.getStatus() == GameStatus.GAME_END_CELEBRATING)) {
+                if (!GameCreator.isInArea(event.getTo(), game.getPos1(), game.getPos2())) {
+                    event.setCancelled(true);
                 }
             }
         }
